+++ zapit/+zapit/+settings/private/checkSettingsAreValid.m
@@ -16,7 +16,6 @@ function [settings,allValid] = checkSettingsAreValid(settings)
     DEFAULT_SETTINGS = default_settings;
     allValid=true;
 
-
     %% 
     %
     % general settings
@@ -66,23 +65,12 @@ function [settings,allValid] = checkSettingsAreValid(settings)
     end
 
 
-    if ~isnumeric(settings.NI.AOrange)
-        fprintf('NI.AOrange should be a number. Setting it to %d \n', ...
-            DEFAULT_SETTINGS.NI.AOrange)
-        settings.NI.AOrange = DEFAULT_SETTINGS.NI.AOrange;
-        allValid=false;
-    elseif settings.NI.AOrange <= 0 
-        fprintf('NI.AOrange should be >0. Setting it to %d \n', ...
-            DEFAULT_SETTINGS.NI.AOrange)
-        settings.NI.AOrange = DEFAULT_SETTINGS.NI.AOrange;
-        allValid=false;
-    end
-
     if ~ischar(settings.NI.triggerChannel)
         fprintf('NI.triggerChannel should be a string. Setting it to "%s"\n', DEFAULT_SETTINGS.NI.triggerChannel)
         settings.NI.triggerChannel = DEFAULT_SETTINGS.NI.triggerChannel;
         allValid=false;
     end
+
 
     if ~ischar(settings.NI.wrapper)
         fprintf('NI.wrapper should be a string. Setting it to "%s"\n', DEFAULT_SETTINGS.NI.wrapper)
@@ -139,15 +127,15 @@ function [settings,allValid] = checkSettingsAreValid(settings)
     %
     % camera
 
-    if ~isnumeric(settings.camera.connection_index)
-        fprintf('camera.connection_index should be a number. Setting it to %d \n', ...
-            DEFAULT_SETTINGS.camera.connection_index)
-        settings.camera.connection_index = DEFAULT_SETTINGS.camera.connection_index;
+    if ~isnumeric(settings.camera.connectionIndex)
+        fprintf('camera.connectionIndex should be a number. Setting it to %d \n', ...
+            DEFAULT_SETTINGS.camera.connectionIndex)
+        settings.camera.connectionIndex = DEFAULT_SETTINGS.camera.connectionIndex;
         allValid=false;
-    elseif settings.camera.connection_index<=0
-        fprintf('camera.connection_index should be >0. Setting it to %d \n', ...
-            DEFAULT_SETTINGS.camera.connection_index)
-        settings.camera.connection_index = DEFAULT_SETTINGS.camera.connection_index;
+    elseif settings.camera.connectionIndex<=0
+        fprintf('camera.connectionIndex should be >0. Setting it to %d \n', ...
+            DEFAULT_SETTINGS.camera.connectionIndex)
+        settings.camera.connectionIndex = DEFAULT_SETTINGS.camera.connectionIndex;
         allValid=false;
     end